Correction de la suppression des commentaires, droits octroyés

+++ app/Http/Controllers/CommentaireController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\Commentaire;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Log;
+
 
 class CommentaireController extends Controller
 {
@@ -36,9 +38,14 @@ class CommentaireController extends Controller
         $user = Auth::user();
         $jeu_id = $commentaire->jeu->id;
         if ($user->can('delete', $commentaire)  || $user->can('isAdmin',$commentaire) )  {
-            $commentaire->delete();
+            Log::info("Delete ok : ");
+            $action = $request->get('action','annule');
+            if ($action=='valide') {
+                $commentaire->delete();
+            }
             return redirect()->route('jeu_show', ['id' => $jeu_id])->with('status', 'Commentaire supprimé');
         } else {
+            Log::info("Delete impossible : ");
             return redirect()->route('jeu_show', ['id' => $jeu_id])->with('status', 'Impossible de supprimer le commentaire');
         }
     }
